fixed intermittent error
Sometimes program would try to dispose resources that GC had already
disposed causing an exception. Removed closing of file from finalizer
as this is what was causing the problem. Instead signal from the reader
that once we read the whole file, we can now close it.

Also commented out “time elapsed” and “memory used” statements so they
won’t be written to the console.